backend/requirements.txt
fastapi
uvicorn
gunicorn
transformers
torch
safetensors
python-dotenv